move functions to script per pr feedback
